requirements.txt
pimoroni-sgp30
smbus2
influxdb-client